Fail by default when encountering interceptor annotations on private methods
