+++ extensions/arc/deployment/src/main/java/io/quarkus/arc/deployment/ArcConfig.java
@@ -77,12 +77,12 @@ public class ArcConfig {
     public boolean transformPrivateInjectedFields;
 
     /**
-     * If set to true, the build fails if a private method that is neither an observer nor a producer, is annotated with an
-     * interceptor binding.
+     * If set to true (the default), the build fails if a private method that is neither an observer nor a producer, is
+     * annotated with an interceptor binding.
      * An example of this is the use of {@code Transactional} on a private method of a bean.
      * If set to false, Quarkus simply logs a warning that the annotation will be ignored.
      */
-    @ConfigItem(defaultValue = "false")
+    @ConfigItem(defaultValue = "true")
     public boolean failOnInterceptedPrivateMethod;
 
     /**
