+++ extensions/arc/deployment/src/test/java/io/quarkus/arc/test/interceptor/IgnoredPrivateInterceptedMethodTest.java
@@ -17,6 +17,7 @@ import jakarta.interceptor.Interceptor;
 import jakarta.interceptor.InterceptorBinding;
 import jakarta.interceptor.InvocationContext;
 
+import org.jboss.shrinkwrap.api.asset.StringAsset;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
 
@@ -27,7 +28,9 @@ public class IgnoredPrivateInterceptedMethodTest {
     @RegisterExtension
     static final QuarkusUnitTest config = new QuarkusUnitTest()
             .withApplicationRoot((jar) -> jar
-                    .addClasses(SimpleBean.class, SimpleInterceptor.class, Simple.class));
+                    .addClasses(SimpleBean.class, SimpleInterceptor.class, Simple.class)
+                    .addAsResource(new StringAsset("quarkus.arc.fail-on-intercepted-private-method=false"),
+                            "application.properties"));
 
     @Inject
     SimpleBean simpleBean;
